Separate file for base_il_trainer, rgb -> bgr fix
