+++ habitat_baselines/__init__.py
@@ -4,8 +4,11 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
+from habitat_baselines.common.base_il_trainer import BaseILTrainer
 from habitat_baselines.common.base_trainer import BaseRLTrainer, BaseTrainer
-from habitat_baselines.il.trainers.edfe_trainer import EDFETrainer
+from habitat_baselines.il.trainers.eqa_cnn_pretrain_trainer import (
+    EQACNNPretrainTrainer,
+)
 from habitat_baselines.il.trainers.eqa_trainer import EQATrainer
 from habitat_baselines.il.trainers.nav_trainer import NavTrainer
 from habitat_baselines.il.trainers.vqa_trainer import VQATrainer
@@ -15,9 +18,10 @@ from habitat_baselines.rl.ppo.ppo_trainer import PPOTrainer, RolloutStorage
 __all__ = [
     "BaseTrainer",
     "BaseRLTrainer",
+    "BaseILTrainer",
     "PPOTrainer",
     "RolloutStorage",
-    "EDFETrainer",
+    "EQACNNPretrainTrainer",
     "VQATrainer",
     "NavTrainer",
     "EQATrainer",
